backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-cloud-documentai==3.5.0
google-cloud-aiplatform==1.103.0
wandb==0.21.0
requests==2.32.4
pydantic==2.11.7
python-multipart==0.0.6 